extract movable-line-editor
